// src/app/@core/types/column.interface.ts
import { NzTableFilterFn, NzTableFilterList, NzTableSortFn, NzTableSortOrder } from 'ng-zorro-antd/table';

export interface ColumnItem {
    name: string;
    sortOrder: NzTableSortOrder | null;
    sortFn: NzTableSortFn| null;
    listOfFilter?: NzTableFilterList | null;
    filterFn?: NzTableFilterFn| null;
    filterMultiple: boolean;
    sortDirections: NzTableSortOrder[];
    colWidth: string;
  }